[BOOK-181] fix: apis - 로그인 시 외부 호출로 인한 트랜잭션 문제 해결
- signIn 로직에서 외부 API 호출과 DB 트랜잭션의 경계를 분리했습니다.
- AuthUseCase.signIn()은 트랜잭션 없이 외부 API 호출 및 흐름 제어만 담당하도록 수정했습니다.
- 실제 DB 작업은 UserSignInService.processSignIn()이라는 별도의 트랜잭션 메서드로 격리하여 처리합니다.
- 외부 API 지연이 DB 커넥션 풀에 영향을 주던 잠재적인 성능 문제를 해결하고, 로그인 프로세스의 안정성을 확보했습니다.
